Update file flags through smbclient's `setmode` command
SMBFile::updateProperties(): update flags on files by calling the `setmode`
command of smbclient. Add a decoder function for the Win32 property string,
because Win32 passes its file flags in that form.

Unfortunately, setting flags only appears to work on files, not directories. On
files, the only flag that appears to be settable is the Hidden flag.