Pass parent reference to child instead of just path
Giving a child a handle to the parent's $this lets the child call the parent's
public methods directly. A child must call the parent's ::invalidate_cache()
method to ask it to delete its disk cache file. Previously we used a function
in the global namespace to get the server tree root, and then a lookup method
on that root based on the parent's path name to get the parent object, but this
way is cleaner.