Auth: add this class to find and hold auth data
Place the auth code into a single class and pass a pointer to that class
around.

Current advantage: this avoids having to pass both username and password
around individually. It also cleans up server.php since all the auth
setup is done elsewhere.

Future advantage: more freedom in how we do auth, It may some day be
possible to define different "views" of the login name for different
tasks. The login to Samba may occur with a certain domain or workgroup