Cache: password also contributes to filename
Filename is unique for function, URI, username and password.
Include password here to avoid decode errors when someone logs in with a
valid username and wrong password. If the password would not contribute
to the filename, we would try to open the existing cache file and fail
to decode it.